Makes v3 keygen play nicely with install action
Two changes here. First, we remove the sys.exit() calls inside the v3
keygen logic. Those worked well enough when the script is invoked
directly via the subcommand, but causes the `./securedrop-admin install`
exit to quit before calling out to Ansible. We replace the sys.exit
calls with `return 0`, since the parent function expects an exit code
to be returned, and passes this to the shell.

Second, we stop calling the keygen logic automatically as part of the
install action. If anything, it'd be more appropriate to configure those
values as part of `sdconfig`, but since the generation depends on the
site config already being evaluted (i.e. v3_onion_services=True), we
shouldn't automatically generate keys before inspecting the config.